Clean out frustum plane code, add test
