+++ modules/core/src/lib/view-manager.js
@@ -24,16 +24,7 @@ import View from '../views/view';
 import Viewport from '../viewports/viewport';
 import log from '../utils/log';
 import {flatten} from '../utils/flatten';
-
-import {Vector3} from 'math.gl';
-
-const cameraPosition = new Vector3();
-const cameraDirection = new Vector3();
-const cameraUp = new Vector3();
-const cameraRight = new Vector3();
-const nearCenter = new Vector3();
-const farCenter = new Vector3();
-const a = new Vector3();
+import {getFrustumPlanes} from '../utils/math-utils';
 
 export default class ViewManager {
   constructor(props = {}) {
@@ -153,75 +144,15 @@ export default class ViewManager {
     const {near, far, fovy} = view.props;
     const aspect = width / height;
 
-    cameraPosition.copy(viewport.cameraPosition);
-    cameraDirection.copy(viewport.cameraDirection).normalize();
-    cameraUp.copy(viewport.cameraUp);
-    cameraRight
-      .copy(cameraDirection)
-      .cross(cameraUp)
-      .normalize();
-    cameraUp.copy(cameraRight).cross(cameraDirection); // Orthogonalize
-
-    const nearHeight = 2 * Math.tan(fovy / 2) * near;
-    const nearWidth = nearHeight * aspect;
-
-    nearCenter
-      .copy(cameraDirection)
-      .scale(near)
-      .add(cameraPosition);
-    farCenter
-      .copy(cameraDirection)
-      .scale(far)
-      .add(cameraPosition);
-
-    const planes = {
-      near: {
-        d: cameraDirection.dot(nearCenter),
-        n: cameraDirection.clone().negate()
-      },
-      far: {
-        d: cameraDirection.dot(farCenter),
-        n: cameraDirection.clone()
-      }
-    };
-
-    a.copy(cameraRight)
-      .scale(nearWidth * 0.5)
-      .add(nearCenter)
-      .subtract(cameraPosition)
-      .normalize();
-    let n = new Vector3(a).cross(cameraUp);
-    let d = cameraPosition.dot(n);
-    planes.right = {n, d};
-
-    a.copy(cameraRight)
-      .scale(-nearWidth * 0.5)
-      .add(nearCenter)
-      .subtract(cameraPosition)
-      .normalize();
-    n = new Vector3(cameraUp).cross(a);
-    d = cameraPosition.dot(n);
-    planes.left = {n, d};
-
-    a.copy(cameraUp)
-      .scale(nearHeight * 0.5)
-      .add(nearCenter)
-      .subtract(cameraPosition)
-      .normalize();
-    n = new Vector3(cameraRight).cross(a);
-    d = cameraPosition.dot(n);
-    planes.top = {n, d};
-
-    a.copy(cameraUp)
-      .scale(-nearHeight * 0.5)
-      .add(nearCenter)
-      .subtract(cameraPosition)
-      .normalize();
-    n = new Vector3(a).cross(cameraRight);
-    d = cameraPosition.dot(n);
-    planes.bottom = {n, d};
-
-    return planes;
+    return getFrustumPlanes({
+      aspect,
+      near,
+      far,
+      fovy,
+      position: viewport.cameraPosition,
+      direction: viewport.cameraDirection,
+      up: viewport.cameraUp
+    });
   }
 
   /**
