+++ modules/core/src/utils/math-utils.js
@@ -1,5 +1,7 @@
 // Extensions to math.gl library. Intended to be folded back.
 
+import {Vector3} from 'math.gl';
+
 // Helper, avoids low-precision 32 bit matrices from gl-matrix mat4.create()
 export function createMat4() {
   return [1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1];
@@ -13,6 +15,90 @@ export function extractCameraVectors({viewMatrix, viewMatrixInverse}) {
     direction: [viewMatrix[2], viewMatrix[6], viewMatrix[10]],
     up: [viewMatrix[1], viewMatrix[5], viewMatrix[9]]
   };
+}
+
+const cameraPosition = new Vector3();
+const cameraDirection = new Vector3();
+const cameraUp = new Vector3();
+const cameraRight = new Vector3();
+const nearCenter = new Vector3();
+const farCenter = new Vector3();
+const a = new Vector3();
+const DEG_TO_RADIAN = Math.PI / 180;
+
+/* eslint-disable max-statements */
+export function getFrustumPlanes({aspect, near, far, fovy, position, direction, up}) {
+  cameraPosition.copy(position);
+  cameraDirection.copy(direction).normalize();
+  cameraUp.copy(up);
+  fovy *= DEG_TO_RADIAN;
+
+  cameraRight
+    .copy(cameraDirection)
+    .cross(cameraUp)
+    .normalize();
+  cameraUp.copy(cameraRight).cross(cameraDirection); // Orthogonalize
+
+  const nearHeight = 2 * Math.tan(fovy / 2) * near;
+  const nearWidth = nearHeight * aspect;
+
+  nearCenter
+    .copy(cameraDirection)
+    .scale(near)
+    .add(cameraPosition);
+  farCenter
+    .copy(cameraDirection)
+    .scale(far)
+    .add(cameraPosition);
+
+  const planes = {
+    near: {
+      d: cameraDirection.dot(nearCenter),
+      n: cameraDirection.clone().negate()
+    },
+    far: {
+      d: cameraDirection.dot(farCenter),
+      n: cameraDirection.clone()
+    }
+  };
+
+  a.copy(cameraRight)
+    .scale(nearWidth * 0.5)
+    .add(nearCenter)
+    .subtract(cameraPosition)
+    .normalize();
+  let n = new Vector3(a).cross(cameraUp);
+  let d = cameraPosition.dot(n);
+  planes.right = {n, d};
+
+  a.copy(cameraRight)
+    .scale(-nearWidth * 0.5)
+    .add(nearCenter)
+    .subtract(cameraPosition)
+    .normalize();
+  n = new Vector3(cameraUp).cross(a);
+  d = cameraPosition.dot(n);
+  planes.left = {n, d};
+
+  a.copy(cameraUp)
+    .scale(nearHeight * 0.5)
+    .add(nearCenter)
+    .subtract(cameraPosition)
+    .normalize();
+  n = new Vector3(cameraRight).cross(a);
+  d = cameraPosition.dot(n);
+  planes.top = {n, d};
+
+  a.copy(cameraUp)
+    .scale(-nearHeight * 0.5)
+    .add(nearCenter)
+    .subtract(cameraPosition)
+    .normalize();
+  n = new Vector3(a).cross(cameraRight);
+  d = cameraPosition.dot(n);
+  planes.bottom = {n, d};
+
+  return planes;
 }
 
 // export function mod(value, divisor) {
